+++ IBLL/DTO/EpisodioDTO.cs
@@ -6,7 +6,7 @@ namespace IBLL.DTO
     {
         public int? episidid { get; set; }
         public int? epispazi { get; set; }
-        public int? epistipo { get; set; }
+        public short? epistipo { get; set; }
         public DateTime? episdain { get; set; }
         public DateTime? episdafi { get; set; }
         public bool? epischiu { get; set; }
